requirements.txt
boto3
Django
django-storages
djangorestframework
pillow
psycopg2-binary
python-decouple
